standalone_modules/logfire_nl_sql_streamlit/requirements.txt
streamlit>=1.36.0
httpx>=0.27.0
pandas>=2.2.2
pyarrow>=16.1.0
python-dotenv>=1.0.1
tenacity>=8.2.3
openai>=1.35.0
